+++ play_pi/api/serializers.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from play_pi import models
+
+
+class TrackSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.Track
+        fields = '__all__'
+
+
+class RadioSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.RadioStation
+        fields = '__all__'
